Refactor a organización de código del generador de reporte
    refs:#238
