+++ series_tiempo_ar_api/libs/indexing/report/report_generator.py
@@ -1,17 +1,14 @@
 #!coding=utf8
 from __future__ import unicode_literals
 
-import json
-
 from django.conf import settings
 from django.contrib.auth.models import Group
 from django.core.mail.message import EmailMultiAlternatives
 from django.template.loader import render_to_string
 from django.utils import timezone
-from pydatajson import DataJson
 
 from series_tiempo_ar_api.apps.management.models import Indicator, Node
-from series_tiempo_ar_api.apps.api.models import Catalog, Dataset, Field, Distribution
+from series_tiempo_ar_api.libs.indexing.report.indicators_generator import IndicatorsGenerator
 from series_tiempo_ar_api.libs.indexing.report.indicators import IndicatorLoader
 
 
@@ -26,8 +23,12 @@ class ReportGenerator(object):
         self.task.finished = timezone.now()
         self.task.status = self.task.FINISHED
         self.task.save()
-        self.persist_indicators()
-        self.calculate_indicators()
+
+        self.indicators_loader.load_indicators_into_db(self.task)
+
+        for node in Node.objects.filter(indexable=True):
+            IndicatorsGenerator(node, self.task).generate()
+
         self.generate_email()
 
         # Reportes de catálogo individual
@@ -47,7 +48,8 @@ class ReportGenerator(object):
             'is_partial_report': bool(node),
         }
         context.update({
-            indicator: self._get_indicator_value(indicator, node=node) for indicator, _ in Indicator.TYPE_CHOICES
+            indicator: self._get_indicator_value(indicator, node=node)
+            for indicator, _ in Indicator.TYPE_CHOICES
         })
         self.send_email(context, node)
 
@@ -90,90 +92,3 @@ class ReportGenerator(object):
             return 0
 
         return int(sum([indic.value for indic in indicator_queryset]))
-
-    def calculate_indicators(self):
-
-        nodes = Node.objects.filter(indexable=True)
-        catalog_ids = nodes.values_list('catalog_id', flat=True)
-        for catalog_model in Catalog.objects.filter(identifier__in=catalog_ids):
-            node = nodes.get(catalog_id=catalog_model.identifier)
-
-            self.calculate_catalog_indicators(node)
-            data_json = DataJson(json.loads(node.catalog))
-
-            self.calculate_series_indicators(node, data_json)
-            self.calculate_distribution_indicators(node, data_json)
-            self.calculate_dataset_indicators(node, data_json)
-
-    def calculate_catalog_indicators(self, node):
-        catalog_model = Catalog.objects.get(identifier=node.catalog_id)
-        updated = catalog_model.updated
-        self.task.indicator_set.create(type=Indicator.CATALOG_UPDATED, value=updated, node=node)
-        not_updated = 1 - updated
-        self.task.indicator_set.create(type=Indicator.CATALOG_NOT_UPDATED, value=not_updated,
-                                       node=node)
-        self.task.indicator_set.create(type=Indicator.CATALOG_TOTAL, value=1, node=node)
-
-    def calculate_series_indicators(self, node, data_json):
-        fields_total = len(data_json.get_fields(only_time_series=True))
-        self.task.indicator_set.create(type=Indicator.FIELD_TOTAL, value=fields_total, node=node)
-
-        catalog = Catalog.objects.get(identifier=node.catalog_id)
-
-        indexable = Field.objects.filter(distribution__dataset__catalog=catalog,
-                                         distribution__dataset__indexable=True).count()
-        self.task.indicator_set.create(type=Indicator.FIELD_INDEXABLE, value=indexable, node=node)
-
-        not_indexable = Field.objects.filter(distribution__dataset__catalog=catalog,
-                                             distribution__dataset__indexable=False,
-                                             distribution__dataset__present=True).count()
-        self.task.indicator_set.create(type=Indicator.FIELD_NOT_INDEXABLE, value=not_indexable, node=node)
-        updated = self.task.indicator_set.get_or_create(type=Indicator.FIELD_UPDATED, node=node)[0].value
-
-        not_updated = indexable - updated
-        self.task.indicator_set.create(type=Indicator.FIELD_NOT_UPDATED, value=not_updated, node=node)
-
-    def calculate_distribution_indicators(self, node, data_json):
-        distribution_total = len(data_json.get_distributions(only_time_series=True))
-        self.task.indicator_set.create(type=Indicator.DISTRIBUTION_TOTAL,
-                                       value=distribution_total,
-                                       node=node)
-
-        catalog = Catalog.objects.get(identifier=node.catalog_id)
-
-        indexable = Distribution.objects.filter(dataset__catalog=catalog,
-                                                dataset__indexable=True).count()
-        self.task.indicator_set.create(type=Indicator.DISTRIBUTION_INDEXABLE, value=indexable, node=node)
-        not_indexable = Distribution.objects.filter(dataset__catalog=catalog,
-                                                    dataset__indexable=False,
-                                                    dataset__present=True).count()
-        self.task.indicator_set.create(type=Indicator.DISTRIBUTION_NOT_INDEXABLE, value=not_indexable, node=node)
-
-        updated = self.task.indicator_set.get_or_create(type=Indicator.DISTRIBUTION_UPDATED, node=node)[0].value
-
-        not_updated = indexable - updated
-        self.task.indicator_set.create(type=Indicator.DISTRIBUTION_NOT_UPDATED, value=not_updated, node=node)
-
-    def calculate_dataset_indicators(self, node, data_json):
-        dataset_total = len(data_json.get_datasets(only_time_series=True))
-        self.task.indicator_set.create(type=Indicator.DATASET_TOTAL, value=dataset_total, node=node)
-
-        catalog = Catalog.objects.get(identifier=node.catalog_id)
-
-        indexable = Dataset.objects.filter(catalog=catalog, indexable=True).count()
-        self.task.indicator_set.create(type=Indicator.DATASET_INDEXABLE, value=indexable, node=node)
-
-        not_indexable = Dataset.objects.filter(catalog=catalog, indexable=False).count()
-        self.task.indicator_set.create(type=Indicator.DATASET_NOT_INDEXABLE, value=not_indexable, node=node)
-
-        updated = Dataset.objects.filter(catalog=catalog, updated=True).count()
-        self.task.indicator_set.create(type=Indicator.DATASET_UPDATED, value=updated, node=node)
-
-        not_updated = indexable - updated
-        self.task.indicator_set.create(type=Indicator.DATASET_NOT_UPDATED, value=not_updated, node=node)
-
-        error = Dataset.objects.filter(catalog=catalog, error=True).count()
-        self.task.indicator_set.create(type=Indicator.DATASET_ERROR, value=error, node=node)
-
-    def persist_indicators(self):
-        self.indicators_loader.load_indicators_into_db(self.task)
